chore: build:ci script to speed up production build for design-system app
